Polish work item 46 implementation
Agent-Logs-Url: https://github.com/pfirmstone/JGDMS/sessions/adc93b75-6aeb-48df-a9ba-ce09c94b0173

+++ JGDMS/jgdms-platform/src/main/java/net/jini/security/policy/DynamicPolicyProvider.java
@@ -705,7 +705,7 @@ Put the policy providers and all referenced classes in the bootstrap class loade
             Class<?> providerClass = Class.forName(inconclusiveLoaderEvictionClass);
             Method evictMethod = providerClass.getMethod(inconclusiveLoaderEvictionMethod);
             Object result = evictMethod.invoke(null);
-            if (result instanceof Integer && ((Integer) result).intValue() > 0
+            if (result instanceof Integer && (Integer) result > 0
                     && logger.isLoggable(Level.FINE)) {
                 logger.log(Level.FINE,
                         "Evicted {0} INCONCLUSIVE preferred proxy classloader(s) after dynamic grant",
